Moved and renamed C structures abstractions
