+++ stmeasures/objects/cstructures.py
